refactor: split report discovery family
